Revert "[ipython] Fix mode name, and remove unmaintainable code"
As discussed at length in #13392, the primary developer of EIN insists
on updating the spacemacs bindings to avoid several-iterations stale commands of
dubious functionality (in particular, multi-worksheet commands).

As the transient state docstring must be "mated" to the updated bindings, he prefers
not having to manually curate it, and exhorts the return of the
automated docstring.  He doesn't insist on it, though, like the above.

bind-map.el counterfeits ein:notebook-mode

bind-map replaces on the `minor-mode-map-alist` the keymap for
ein:notebook-mode with the evilified map, thus making the original
ein:notebook-mode-map inaccessible.

This is might be intentional as one wonders whether having active both the
original keymap and the evilified keymap makes sense.

In EIN's case, yes, we really want the original keymap, e.g., C-c C-c
to execute, to work.  Otherwise C-c C-c defaults to running a python
shell via the python-mode-map, which is really not what we want.