Make spacemacs/cycle-spacemacs-theme work with themes with keywords
Themes can now contain keywords in addition to just the theme name like

   dotspacemacs-themes '((name :location (recipe :fetcher github
                                                    :repo "user/repo"))
                         solarized-light
                         solarized-dark)

However, for theme cycling, we first try to detect whether the current
theme (spacemacs--cur-theme) is in the list of cycleable themes and go on from
there. Since spacemacs--cur-theme is just a name, this fails when
spacemacs--cur-theme is the name of a theme that's specified with additional
keywords in dotspacemacs-themes.

To fix this, check if spacemacs--cur-theme is in the list of theme names.