[factor] Support running graphical Factor listener from Spacemacs
Fuel mode is responsible to connect to a Factor instance.  This change extends
the Factor layer to handle a graphical listener process, to which fuel can
connect afterwards.

A major motivation is also to make it easier to develop with different Factor
versions, which can be specified with (project-/directory-specific) variables.
When starting a Factor listener in a certain location this way, the elisp code
for fuel/factor mode is reloaded from that location.