[go] Remove duplicated code and run golangci-lint only if no errors are found
Before the decision which linter to run before golangci-lint has been
taken by the layer. However this can also directly be done by the
checkers.

Also golangci-lint did run even when the previous checker still reported
errors. With this it is only called if the previous checker is satisfied.

All thanks goes to Tommi Komulainen.