[core] Make spacemacs|use-package-add-hook emit code for warning generation
In my first commit I did not add the warning message properly.
Basically the macro should emit code to create a warning if
the generated code is run and the predicate is not fulfilled.

However in my last commit it did run the code during macro
expansion which would do the testing there which is not
what I wanted.