[vagrant] Correctly bind `vagrant-tramp-term`
According to the README, `SPC a t v t` should start `vagrant-tramp-term`, but
instead, it seems to have been bound to `SPC a t v`.  This means that it
overrode all of the other vagrant keybindings under the prefix `SPC a t v`.

The issue seems to have been a typo from commit e38c33f.

Relates to #13503.