[ipython] Fix bindinding for RET in transient state
In the documentation, it says this performs execute cell and goto next, while
the code was just executing the cell.

This desync could be fixed in either direction, but executing and going to next
feels more ergonomic while in the transient state, since we're jumping around
running rather than editing in that mode, so I've fixed it to do that.