[clojure] Add workaround for corrupted sayid.elc file
It seems to happen from time to time that sayid is not
properly compiled after the installation. In this case
the variable `sayid-version` is `nil` causing
nRepl to fail.

This commit tries to work around this problem by
reinstalling the package if this corruption is detected.

This normally fixes the issue described in #13906.