[org] SPC m s N -> SPC m s w to match mnemonics of SPC n w
I chose to go this way rather than the other way because:

* Whenever you can avoid holding down a modifier key, great!
* It seems like the more specific should yield to the more general

This may be a controversial change, because it could break some folks' finger
memory. However I consider this a very unpleasant inconsistency that inhibits
the development of finger memory, so I thought I would at least suggest it as
the default for Spacemacs.