Move lsp-ui back to be enabled by default
I don't think that setting this to a fixed value
will solve the issue. Right now we have a conflict
between feature rich and simple here.

This is similar to what we have with the search
integration (helm vs ivy) and in the tree
implementation (treemacs vs neotree).

For these we have a workflow during the initial
dotfile creation. Which we need to follow here
too.