Revert "python: replaced local package `pylookup` with the one on github"
This reverts commit 6e0f6501017552ad6213d3ab0328db39c7158faa.
The reason for this is not that the dynamically build pylookup.py
is not found by emacs, this I have already solved locally by checking
the quelpa build directory.

However the issue is that the library is not compatible with python 3
the local version however is.

So this is effectively a fork and must be kept till python 3
compatibility for pylookup has been achieved.

Should fix #14544