Fix expected type of: dotspacemacs-which-key-delay
problem:
Setting dotspacemacs-which-key-delay to an integer, for example: 4

Makes the following message to appear on startup:
(Spacemacs) Error in dotspacemacs/init: Variable: "dotspacemacs-which-key-delay" has value: "4" that doesn’t match its type: "(float)". Validator message: "(user-error Looking for `(float)' in `4' failed because:
  not a float)"

cause:
dotspacemacs-which-key-delay expects the type: float
The float type has the description:
The value must be floating point.

solution:
Change the type to: number
The number type has the description:
The value must be a number (floating point or integer).

Source: info (`C-h i`)
In the section: Elisp > Customization Types > Simple Types