[core] Add :timeout/:idle for spacemacs|define-transient-state (#14742)
* Add :timeout for spacemacs|define-transient-state

This add a :timeout keyword for spacemacs|define-transient-state, it
make use of :timeout keyword defhydra according to
https://github.com/abo-abo/hydra/wiki/internals.

"The :timeout key starts a timer for the corresponding amount of seconds that disables the hydra. Calling any head will refresh the timer."

* Add idle for spacemacs|define-transient-state.

According to https://github.com/abo-abo/hydra/wiki/internals:
This key can delay the appearance of the hint.
Add :timeout and :idle in defhydra body to make it effect.