Fix (activate) company-emoji completion
Currently company-emoji does not get activated by the emoji layer.
Also when it is activated, then the emoji short-name does not get completed to
unicode if `company-emoji-insert-unicode` is `t`.

This commit fixes both issues, by activating company-emoji for text-modes and
advising the emoji inert function to optionally do company-emoji completion.