Remove helm-gitignore from git layer.
Related to #15117, this removes the helm-gitignore from the git layer because it
hasn't been updated to use the new git-modes package. The helm-gitignore
functionality isn't particularly useful anyway, it's just a utility to help
generate .gitignore files.

This change can be reverted if the upstream helm-gitignore package is fixed and
people find the functionality it provides to be useful.